Make 3DGS implementation and example work initially

+++ 3rdparty/spz/splat-types.h
@@ -181,7 +181,7 @@ struct GaussianCloud {
     }
     std::sort(scaleSums.begin(), scaleSums.end());
     float median = scaleSums[(int32_t)(scaleSums.size() / 2)];
-    return (M_PI * 4 / 3) * exp(median);
+    return (float)(M_PI * 4 / 3) * exp(median);
   }
 };
